reorganize the files using headers and libs folders
