+++ .gitattributes
@@ -1,5 +1,5 @@
 # to not include externally sourced libraries in github language metrics
-sqlite3.c linguist-vendored
-sqlite3.h linguist-vendored
-mongoose.c linguist-vendored
-mongoose.h linguist-vendored
+"**/sqlite3.c" linguist-vendored
+"**/sqlite3.h" linguist-vendored
+"**/mongoose.c" linguist-vendored
+"**/mongoose.h" linguist-vendored
