+++ headers/sql_handler.h
@@ -3,7 +3,7 @@
 
 #include <stdio.h>
 #include <string.h>
-#include "sqlite3.h"
+#include "../libs/sqlite3.h"
 
 #define MAX_RESPONSE_SIZE 99000
