+++ libs/sqlite3.c
