+++ libs/sqlite3.h
