+++ sql_handler.c
@@ -1,4 +1,4 @@
-#include "sql_handler.h"
+#include "headers/sql_handler.h"
 
 // execute a select query and return the JSON response
 // (the address of a buffer is passed as a parameter to which the response will be written)
